fullpush

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -10,10 +10,10 @@
  */
 void hbd(unsigned long int n, unsigned long int i)
 {
-	if ((1ul << i) > n)
+	if (((unsigned long int)1 << i) > n)
 		return;
 	hbd(n, i + 1);
-	_putchar('0' + ((n & (1ul << i)) ? 1ul : 0ul));
+	_putchar('0' + ((n & ((unsigned long int)1 << i)) ? 1ul : 0ul));
 }
 /**
  * print_binary - hi
